Almost quadratic growth for the generation's size
We have choosen to try to avoid a linear growth of the generation's size because the counter-example's research was becoming a bruteforce search.
So the "ForAll" now generates a sequence of sizes (not in a lazy way but to be a problem the user must ask for a huge maxSize, in that case we can easily fill up all the available memory, but I think it can be a optimization that we can achieve in a future version, maybe with the generators available in PHP 5.5) which respects the maxSize set by the user or the default one (which is 200).

Quickcheck then cycles on the sequence of the sizes and picks the generation's size for the current evaluation.

In this commit we are switching from a linear growth of the size to an almost quadratic growth using the "triangular number algorithm".

To be clear...
Our size sequence was:
[0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, ...]
and now it is:
[0, 1, 3, 6, 10, 15, 21, 28, 36, 45, ...].