Tuning test because the size growth has changed
This test makes expectations on the duration of a property evaluation, since (now) the size grows in a controlled manner it takes more time to find the counter-example but this is a good thing because previously the size was completely random, resulting in poor shrinking performances.

After some reasoning it is expected that it takes 3 seconds more than before because of the behaviour of the triangle number algorithm (and because the test sleeps for a second for each concatenation) which asks for 4 sizes before asking for a size which causes the bug.