Fixing how Generator\Sequence uses the size
Before this commit it was generating sequences with the length exactly equal to the generation's size.
This is not good, because since the size now is growing quadratically we risk to test only a small amount of sequence lengths, so now it generates sequences with the length less than or equal to the size.